Separate modules for model and plugin.
