+++ codegen-plugin/src/main/kotlin/io/spine/protodata/hello/CodeGenPlugin.kt
