+++ model/src/main/proto/spine/protodata/hello/echo.proto
